refactor(lua): move LuaRuntime to its own module, decouple from Commands
The interpreter no longer knows about Bevy. apply(&mut Commands) becomes
drain_outbox() -> (Vec<Command>, Vec<(String, f32)>) and dispatch_with_world
becomes dispatch_with_query, taking the world-reading extract callback the
systems now build themselves. runtime.rs imports nothing from bevy.

Every existing test moves across unchanged, including the at-most-once
extraction guarantee; one test is added for the new drain contract.
